Initial commit of a HuJSON AST parser and packer
This package provides a HuJSON parser into an AST represented
by the following Go types:
* Value: A JSON literal, object, or array with surrounding whitespace/comments.
* ValueTrimmed: A JSON value without surrounding whitespace/comments.
* Extra: Raw bytes of whitespace and comments.
* Literal: Raw bytes of a JSON literal (i.e., null, boolean, string, or number).
* Object: A JSON object.
* Array: A JSON array.

The Parse function parses raw HuJSON input as a Value.
The Value.Pack method serializes the AST as raw HuJSON output.

The intention is to delete the fork of "encoding/json"
currently here once we migrate all usages to the new code.