feat(hujsonfmt): add basic hujsonfmt CLI command
Added as a separate Go module with its own go.mod file to avoid
polluting the root module's dependency tree.

The CLI interface and flags are very similar to gofmt:

    $ hujsonfmt -h
    usage: hujsonfmt [flags] [path ...]
      -d	display diffs instead of rewriting files
      -l	list files whose formatting differs from hujsonfmt's
      -m	minify results
      -s	standardize results to plain JSON
      -w	write result to (source) file instead of stdout

Given paths can either point directly to a file, or to a directory. In
the case of a directory, it will be recursively walked finding
all *.hujson files.

Input can also be provided via stdin if no paths are provided, or a
single path of "-" is provided:

    $ echo '{\n// hai\n"foo":"bar"}' | ./hujsonfmt
    {
    	// hai
    	"foo": "bar",
    }